Struct circulate::Relay

pub struct Relay { /* private fields */ }

Manages subscriptions and notifications for PubSub.

Implementations

impl Relay

pub fn create_subscriber(&self) -> Subscriber

Create a new Subscriber for this relay.

pub fn publish<Topic: Serialize, P: Serialize>( &self, topic: &Topic, payload: &P ) -> Result<(), Error>

Publishes a payload to all subscribers of topic.

Errors

Returns an error if topic or payload fails to serialize with pot.

pub fn publish_raw<Topic: Into<OwnedBytes>, Payload: Into<OwnedBytes>>( &self, topic: Topic, payload: Payload )

Publishes a payload to all subscribers of topic.

pub fn publish_to_all<'topics, Topics: IntoIterator<Item = &'topics Topic> + 'topics, Topic: Serialize + 'topics, Payload: Serialize>( &self, topics: Topics, payload: &Payload ) -> Result<(), Error>

Publishes a payload to all subscribers of topic.

Errors

Returns an error if topics or payload fail to serialize with pot.

pub fn publish_raw_to_all( &self, topics: impl IntoIterator<Item = OwnedBytes>, payload: impl Into<OwnedBytes> )

Publishes a payload to all subscribers of topic.

pub fn publish_message(&self, message: &Message)

Publishes a message to all subscribers of its topic.